+++ demos/chinook/src/main/java/module-info.java
@@ -28,7 +28,7 @@ module is.codion.framework.demos.chinook {
 	requires is.codion.swing.framework.ui;
 	requires is.codion.plugin.jasperreports;
 	requires is.codion.plugin.imagepanel;
-	requires jasperreports;
+	requires net.sf.jasperreports.core;
 	requires com.formdev.flatlaf.intellijthemes;
 	requires org.kordamp.ikonli.foundation;