+++ demos/manual/src/main/java/module-info.java
@@ -21,7 +21,7 @@
  */
 module is.codion.framework.demos.manual {
 	requires java.desktop;
-	requires jasperreports;
+	requires net.sf.jasperreports.core;
 	requires is.codion.common.core;
 	requires is.codion.dbms.h2;
 	requires is.codion.framework.db.local;